+++ src/app/core/services/constants.service.ts
@@ -1,0 +1,6 @@
+import { InjectionToken } from '@angular/core';
+import { Constants } from '../shared/constants.interface';
+
+export const constantsToken = new InjectionToken<Constants>('Constants');
+
+export const constantsInstance: Constants = { app: 'TaskManager', ver: '1.0' };