+++ src/app/first/components/first/first.component.ts
@@ -1,12 +1,17 @@
 import { Component, Inject, OnInit, Optional } from '@angular/core';
 import { ConfigOptionsService } from 'src/app/core/services/config-options.service';
 import { constantsToken } from 'src/app/core/services/constants.service';
+import { generatedString, GeneratorFactory, GeneratorService } from 'src/app/core/services/generator.service';
 import { Constants } from 'src/app/core/shared/constants.interface';
 
 @Component({
   selector: 'app-first',
   templateUrl: './first.component.html',
-  styleUrls: ['./first.component.scss']
+  styleUrls: ['./first.component.scss'],
+  providers: [
+    GeneratorService,
+    { provide: generatedString, useFactory: GeneratorFactory(10), deps: [GeneratorService] }
+  ]
 })
 export class FirstComponent implements OnInit {
   title: string;
@@ -14,8 +19,9 @@ export class FirstComponent implements OnInit {
   user = 'no user';
 
   constructor(
+    @Inject(generatedString) public randomString10: string,
     @Optional() private configOptionsService: ConfigOptionsService,
-    @Inject(constantsToken) private constants: Constants
+    @Inject(constantsToken) private constants: Constants,
   ) { }
 
   ngOnInit(): void {